Add tests asserting that versions view and versions list hide the Wrangler banner when --json is passed
