+++ packages/wrangler/src/__tests__/vitest.setup.ts
@@ -83,6 +83,7 @@ vi.mock("undici", async (importOriginal) => {
 vi.mock("../package-manager");
 
 vi.mock("../update-check");
+vi.mock("../wrangler-banner");
 
 beforeAll(() => {
 	msw.listen({
